update file organization
