+++ torchaudio/csrc/decoder/src/decoder/LexiconDecoder.h
@@ -9,65 +9,78 @@
 
 #include <unordered_map>
 
-#include "torchaudio/csrc/decoder/src/text/decoder/Decoder.h"
-#include "torchaudio/csrc/decoder/src/text/decoder/lm/LM.h"
+#include "torchaudio/csrc/decoder/src/decoder/Decoder.h"
+#include "torchaudio/csrc/decoder/src/decoder/Trie.h"
+#include "torchaudio/csrc/decoder/src/decoder/lm/LM.h"
 
 namespace torchaudio {
 namespace lib {
 namespace text {
 
-struct LexiconFreeDecoderOptions {
+struct LexiconDecoderOptions {
   int beamSize; // Maximum number of hypothesis we hold after each step
   int beamSizeToken; // Maximum number of tokens we consider at each step
   double beamThreshold; // Threshold to prune hypothesis
   double lmWeight; // Weight of lm
+  double wordScore; // Word insertion score
+  double unkScore; // Unknown word insertion score
   double silScore; // Silence insertion score
-  bool logAdd;
+  bool logAdd; // If or not use logadd when merging hypothesis
   CriterionType criterionType; // CTC or ASG
 };
 
 /**
- * LexiconFreeDecoderState stores information for each hypothesis in the beam.
+ * LexiconDecoderState stores information for each hypothesis in the beam.
  */
-struct LexiconFreeDecoderState {
+struct LexiconDecoderState {
   double score; // Accumulated total score so far
   LMStatePtr lmState; // Language model state
-  const LexiconFreeDecoderState* parent; // Parent hypothesis
+  const TrieNode* lex; // Trie node in the lexicon
+  const LexiconDecoderState* parent; // Parent hypothesis
   int token; // Label of token
+  int word; // Label of word (-1 if incomplete)
   bool prevBlank; // If previous hypothesis is blank (for CTC only)
 
   double amScore; // Accumulated AM score so far
   double lmScore; // Accumulated LM score so far
 
-  LexiconFreeDecoderState(
+  LexiconDecoderState(
       const double score,
       const LMStatePtr& lmState,
-      const LexiconFreeDecoderState* parent,
+      const TrieNode* lex,
+      const LexiconDecoderState* parent,
       const int token,
+      const int word,
       const bool prevBlank = false,
       const double amScore = 0,
       const double lmScore = 0)
       : score(score),
         lmState(lmState),
+        lex(lex),
         parent(parent),
         token(token),
+        word(word),
         prevBlank(prevBlank),
         amScore(amScore),
         lmScore(lmScore) {}
 
-  LexiconFreeDecoderState()
-      : score(0),
+  LexiconDecoderState()
+      : score(0.),
         lmState(nullptr),
+        lex(nullptr),
         parent(nullptr),
         token(-1),
+        word(-1),
         prevBlank(false),
         amScore(0.),
         lmScore(0.) {}
 
-  int compareNoScoreStates(const LexiconFreeDecoderState* node) const {
+  int compareNoScoreStates(const LexiconDecoderState* node) const {
     int lmCmp = lmState->compare(node->lmState);
     if (lmCmp != 0) {
       return lmCmp > 0 ? 1 : -1;
+    } else if (lex != node->lex) {
+      return lex > node->lex ? 1 : -1;
     } else if (token != node->token) {
       return token > node->token ? 1 : -1;
     } else if (prevBlank != node->prevBlank) {
@@ -77,11 +90,11 @@ struct LexiconFreeDecoderState {
   }
 
   int getWord() const {
-    return -1;
+    return word;
   }
 
   bool isComplete() const {
-    return true;
+    return !parent || parent->word >= 0;
   }
 };
 
@@ -89,26 +102,34 @@ struct LexiconFreeDecoderState {
  * Decoder implements a beam seach decoder that finds the word transcription
  * W maximizing:
  *
- * AM(W) + lmWeight_ * log(P_{lm}(W)) + silScore_ * |{i| pi_i = <sil>}|
+ * AM(W) + lmWeight_ * log(P_{lm}(W)) + wordScore_ * |W_known| + unkScore_ *
+ * |W_unknown| + silScore_ * |{i| pi_i = <sil>}|
  *
  * where P_{lm}(W) is the language model score, pi_i is the value for the i-th
  * frame in the path leading to W and AM(W) is the (unnormalized) acoustic model
- * score of the transcription W. We are allowed to generate words from all the
- * possible combination of tokens.
+ * score of the transcription W. Note that the lexicon is used to limit the
+ * search space and all candidate words are generated from it if unkScore is
+ * -inf, otherwise <UNK> will be generated for OOVs.
  */
-class LexiconFreeDecoder : public Decoder {
+class LexiconDecoder : public Decoder {
  public:
-  LexiconFreeDecoder(
-      LexiconFreeDecoderOptions opt,
+  LexiconDecoder(
+      LexiconDecoderOptions opt,
+      const TriePtr& lexicon,
       const LMPtr& lm,
       const int sil,
       const int blank,
-      const std::vector<float>& transitions)
+      const int unk,
+      const std::vector<float>& transitions,
+      const bool isLmToken)
       : opt_(std::move(opt)),
+        lexicon_(lexicon),
         lm_(lm),
-        transitions_(transitions),
         sil_(sil),
-        blank_(blank) {}
+        blank_(blank),
+        unk_(unk),
+        transitions_(transitions),
+        isLmToken_(isLmToken) {}
 
   void decodeBegin() override;
 
@@ -127,29 +148,35 @@ class LexiconFreeDecoder : public Decoder {
   std::vector<DecodeResult> getAllFinalHypothesis() const override;
 
  protected:
-  LexiconFreeDecoderOptions opt_;
+  LexiconDecoderOptions opt_;
+  // Lexicon trie to restrict beam-search decoder
+  TriePtr lexicon_;
   LMPtr lm_;
+  // Index of silence label
+  int sil_;
+  // Index of blank label (for CTC)
+  int blank_;
+  // Index of unknown word
+  int unk_;
+  // matrix of transitions (for ASG criterion)
   std::vector<float> transitions_;
+  // if LM is token-level (operates on the same level as acoustic model)
+  // or it is word-level (in case of false)
+  bool isLmToken_;
 
   // All the hypothesis new candidates (can be larger than beamsize) proposed
   // based on the ones from previous frame
-  std::vector<LexiconFreeDecoderState> candidates_;
+  std::vector<LexiconDecoderState> candidates_;
 
   // This vector is designed for efficient sorting and merging the candidates_,
   // so instead of moving around objects, we only need to sort pointers
-  std::vector<LexiconFreeDecoderState*> candidatePtrs_;
+  std::vector<LexiconDecoderState*> candidatePtrs_;
 
   // Best candidate score of current frame
   double candidatesBestScore_;
 
-  // Index of silence label
-  int sil_;
-
-  // Index of blank label (for CTC)
-  int blank_;
-
   // Vector of hypothesis for all the frames so far
-  std::unordered_map<int, std::vector<LexiconFreeDecoderState>> hyp_;
+  std::unordered_map<int, std::vector<LexiconDecoderState>> hyp_;
 
   // These 2 variables are used for online decoding, for hypothesis pruning
   int nDecodedFrames_; // Total number of decoded frames.
