+++ torchaudio/csrc/decoder/src/decoder/Trie.h
