+++ torchaudio/csrc/decoder/src/decoder/Utils.h
@@ -12,7 +12,7 @@
 #include <unordered_map>
 #include <vector>
 
-#include "torchaudio/csrc/decoder/src/text/decoder/lm/LM.h"
+#include "torchaudio/csrc/decoder/src/decoder/lm/LM.h"
 
 namespace torchaudio {
 namespace lib {
