+++ torchaudio/csrc/decoder/src/dictionary/Dictionary.cpp
@@ -8,10 +8,10 @@
 #include <iostream>
 #include <stdexcept>
 
-#include "torchaudio/csrc/decoder/src/common/String.h"
-#include "torchaudio/csrc/decoder/src/common/System.h"
-#include "torchaudio/csrc/decoder/src/text/dictionary/Dictionary.h"
-#include "torchaudio/csrc/decoder/src/text/dictionary/Utils.h"
+#include "torchaudio/csrc/decoder/src/dictionary/String.h"
+#include "torchaudio/csrc/decoder/src/dictionary/System.h"
+#include "torchaudio/csrc/decoder/src/dictionary/Dictionary.h"
+#include "torchaudio/csrc/decoder/src/dictionary/Utils.h"
 
 namespace torchaudio {
 namespace lib {
